auth refactor done
