--- client/src/components/Authentication/CenteredContainer.jsx
@@ -1,14 +1,0 @@
-import { Container } from 'react-bootstrap';
-
-export default function CenteredContainer({ children }) {
-  return (
-    <Container
-      className="d-flex align-items-center justify-content-center"
-      style={{ minHeight: "100vh" }}
-    >
-      <div className="w-100" style={{ maxWidth: "400px" }}>
-        {children}
-      </div>
-    </Container>
-  )
-}
