--- client/src/components/Authentication/Login.jsx
@@ -1,67 +1,0 @@
-import { useRef, useState } from 'react';
-import { Form, Button, Card, Alert } from 'react-bootstrap';
-import { useAuth } from '../../contexts/AuthContext';
-import { Link, useNavigate } from 'react-router-dom';
-import CenteredContainer from './CenteredContainer';
-
-export default function Login() {
-  const emailRef = useRef();
-  const userNameRef = useRef();
-  const passwordRef = useRef();
-  const { login } = useAuth();
-  const [error, setError] = useState('');
-  const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
-
-  async function handleSubmit(e) {
-    e.preventDefault();
-
-    try {
-      setError('');
-      setLoading(true);;
-      await login(emailRef.current.value, userNameRef.current.value, passwordRef.current.value);
-      navigate("/");
-    } catch {
-      setError("Failed to log in");
-    }
-
-    setLoading(false);
-  }
-
-  return (
-    <div className='bg-light'>
-      
-    <CenteredContainer>
-      <Card>
-        <Card.Body>
-          <h2 className="text-center mb-4">Log In</h2>
-          {error && <Alert variant="danger">{error}</Alert>}
-          <Form onSubmit={handleSubmit}>
-            <Form.Group id="email">
-              <Form.Label>Email</Form.Label>
-              <Form.Control type="email" ref={emailRef} required />
-            </Form.Group>
-            <Form.Group id="username">
-              <Form.Label>User Name</Form.Label>
-              <Form.Control type="username" ref={userNameRef} required />
-            </Form.Group>
-            <Form.Group id="password">
-              <Form.Label>Password</Form.Label>
-              <Form.Control type="password" ref={passwordRef} required />
-            </Form.Group>
-            <Button disabled={loading} className="w-100 mt-4" type="submit">
-              Log In
-            </Button>
-          </Form>
-          <div className="w-100 text-center mt-3">
-            <Link to="/forgot-password">Forgot Password?</Link>
-          </div>
-        </Card.Body>
-      </Card>
-      <div className="w-100 text-center mt-2">
-        Create an account? <Link to="/signup">Sign Up</Link>
-      </div>
-    </CenteredContainer>
-    </div>
-  )
-}
