+++ client/src/components/Authentication/PrivateRoute.tsx
@@ -1,8 +1,8 @@
 import { Navigate } from "react-router-dom";
 import { useAuth } from "../../contexts/AuthContext";
 
-export default function PrivateRoute({ children }) {
-  
+export default function PrivateRoute({ children }: { children: React.ReactNode }) {
+
   const { Currentuser } = useAuth();
 
   console.log("user:", Currentuser);
