+++ client/src/contexts/AuthContext.tsx
@@ -1,21 +1,23 @@
 import React, { useContext, useState, useEffect } from "react"
 import { auth } from "../firebase"
+import { LoginType } from "../Type";
 
-const AuthContext = React.createContext()
+const AuthContext = React.createContext<LoginType | null>(null);
+;
 
 export function useAuth() {
   return useContext(AuthContext)
 }
 
-export function AuthProvider({ children }) {
-  const [currentUser, setCurrentUser] = useState()
+export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const [currentUser, setCurrentUser] = useState() as any;
   const [loading, setLoading] = useState(true)
 
-  function signup(email, password) {
+  function signup(email: string, password: string) {
     return auth.createUserWithEmailAndPassword(email, password)
   }
 
-  function login(email, password) {
+  function login(email: string, password: string) {
     return auth.signInWithEmailAndPassword(email, password)
   }
 
@@ -23,15 +25,15 @@ export function AuthProvider({ children }) {
     return auth.signOut()
   }
 
-  function resetPassword(email) {
+  function resetPassword(email: string) {
     return auth.sendPasswordResetEmail(email)
   }
 
-  function updateEmail(email) {
+  function updateEmail(email: string) {
     return currentUser.updateEmail(email)
   }
 
-  function updatePassword(password) {
+  function updatePassword(password: string) {
     return currentUser.updatePassword(password)
   }
 
@@ -44,18 +46,12 @@ export function AuthProvider({ children }) {
     return unsubscribe
   }, [])
 
-  const value = {
-    currentUser,
-    login,
-    signup,
-    logout,
-    resetPassword,
-    updateEmail,
-    updatePassword
-  }
+  // const value = {
+
+  // }
 
   return (
-    <AuthContext.Provider value={value}>
+    <AuthContext.Provider value={ }>
       {!loading && children}
     </AuthContext.Provider>
   )
