--- client/src/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import 'bootstrap/dist/css/bootstrap.min.css';
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  
-    <App />
-);
